init useProfileSearch hook
